requirements.txt
python-telegram-bot==21.5
google-generativeai==0.8.3
apscheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.2
